feat: migrate from PostgreSQL to SQLite
Replace PostgreSQL/pgx with SQLite (modernc.org/sqlite) for a simpler
local-only setup — no Docker, no background process, just a single file.

- Rewrite SQL migrations and queries to SQLite dialect
- Switch sqlc engine from postgresql to sqlite
- Replace pgxpool.Pool with database/sql (*sql.DB)
- Convert pgx.Batch usage to database/sql transactions
- Update all int32/int16/bool types to int64 (SQLite integer type)
- Remove docker-compose.yml, update Makefile and config

+++ internal/handler/dashboard.go
@@ -11,7 +11,7 @@ import (
 
 type DashboardStore interface {
 	GetGameState(ctx context.Context) (generated.GameState, error)
-	ListTeamMembers(ctx context.Context, gameStateID int32) ([]generated.ListTeamMembersRow, error)
+	ListTeamMembers(ctx context.Context, gameStateID int64) ([]generated.ListTeamMembersRow, error)
 }
 
 type DashboardHandler struct {
